feat(routines): show local human-readable time alongside raw timestamps
Run-history API responses (RunSummary/FleetRunSummary), the daemon's structured
JSON log, and the UI's relative-time displays now also expose an absolute,
human-readable local-time form next to the existing raw Unix timestamp /
relative "N ago" text, so timestamps are readable without doing epoch math.

Split svc_list_runs/svc_list_all_runs out of service_trigger.rs into a new
service_runs.rs to stay under the 500-line-per-file gate.

Closes #1159
