+++ apis/openapi.json
@@ -1235,6 +1235,7 @@
           "routine_title",
           "workbench",
           "started_at",
+          "started_at_local",
           "status"
         ],
         "properties": {
@@ -1255,6 +1256,13 @@
             "description": "Unix seconds the run finished (`exit_code` file's mtime), `None` while running or unknown.",
             "minimum": 0
           },
+          "finished_at_local": {
+            "type": [
+              "string",
+              "null"
+            ],
+            "description": "`finished_at` as a human-readable local timestamp, when finished."
+          },
           "routine_id": {
             "type": "string",
             "description": "The routine this run belongs to."
@@ -1268,6 +1276,10 @@
             "format": "int64",
             "description": "Unix seconds the run was triggered.",
             "minimum": 0
+          },
+          "started_at_local": {
+            "type": "string",
+            "description": "`started_at` as a human-readable local (daemon machine's timezone) timestamp."
           },
           "status": {
             "$ref": "#/components/schemas/RunStatus",
@@ -1672,6 +1684,7 @@
         "required": [
           "workbench",
           "started_at",
+          "started_at_local",
           "status"
         ],
         "properties": {
@@ -1692,6 +1705,13 @@
             "description": "Unix seconds the run finished (`exit_code` file's mtime), `None` while running or unknown.",
             "minimum": 0
           },
+          "finished_at_local": {
+            "type": [
+              "string",
+              "null"
+            ],
+            "description": "`finished_at` as a human-readable local timestamp, when finished."
+          },
           "retention_expires_at": {
             "type": [
               "integer",
@@ -1706,6 +1726,10 @@
             "format": "int64",
             "description": "Unix seconds the run was triggered.",
             "minimum": 0
+          },
+          "started_at_local": {
+            "type": "string",
+            "description": "`started_at` as a human-readable local (daemon machine's timezone) timestamp."
           },
           "status": {
             "$ref": "#/components/schemas/RunStatus",
