+++ ui/src/cron_utils_tests.rs
@@ -1,10 +1,13 @@
 //! Host-side unit tests for the pure cron helpers in [`super`]: `parse_cron`'s
-//! field-count normalization and `describe_cron_live`'s validity/description
-//! pairing. `reltime` is excluded — it calls `js_sys::Date::now()` and needs a
-//! wasm/DOM host, so it isn't host-testable (mirrors the `refresh.rs`/
-//! `schedule.rs` split between pure and DOM-touching logic).
+//! field-count normalization, `describe_cron_live`'s validity/description
+//! pairing, and `abstime`'s absolute-timestamp formatting. `reltime` is
+//! excluded — it calls `js_sys::Date::now()` and needs a wasm/DOM host, so
+//! it isn't host-testable (mirrors the `refresh.rs`/`schedule.rs` split
+//! between pure and DOM-touching logic). `abstime` takes its instant as a
+//! plain `u64` rather than reading the clock, so it stays host-testable.
 
 use super::*;
+use chrono::TimeZone;
 
 #[test]
 fn parse_cron_rejects_empty_and_blank() {
@@ -63,4 +66,22 @@ fn describe_cron_live_describes_valid_expression() {
     let (valid, description) = describe_cron_live("0 * * * *");
     assert!(valid);
     assert!(!description.is_empty());
+}
+
+#[test]
+fn abstime_zero_is_a_placeholder() {
+    assert_eq!(abstime(0), "—");
+}
+
+#[test]
+fn abstime_formats_a_known_instant() {
+    // Built in `Local` directly so the expected components match `abstime`'s output regardless
+    // of the host's timezone.
+    let dt = Local.with_ymd_and_hms(2026, 6, 21, 12, 0, 30).unwrap();
+    assert_eq!(abstime(dt.timestamp() as u64), "Jun 21, 2026 12:00");
+}
+
+#[test]
+fn abstime_falls_back_to_dash_for_an_out_of_range_instant() {
+    assert_eq!(abstime(i64::MAX as u64), "—");
 }
