+++ ui/src/main.rs
@@ -26,7 +26,7 @@ mod schedule_heatmap_grid;
 mod settings;
 mod shell_dialogs;
 use command_palette::CommandPalette;
-pub(crate) use cron_utils::{describe_cron_live, parse_cron, reltime};
+pub(crate) use cron_utils::{abstime, describe_cron_live, parse_cron, reltime};
 use header::Header;
 pub(crate) use health::{Health, Toast, ToastKind};
 use overview::OverviewPage;
